PRSD-NONE: move all path segments to UrlSegmentConstants.kt
